+++ m/scripts/publish.sh
@@ -14,3 +14,6 @@ python3 -m twine upload .stage-pypi/dist/*
 
 # Let github know we released
 m github release --owner jmlopez-rod --repo m --version "$M_TAG"
+
+# Create v tags
+m git tag_release --version "$M_TAG"